implemented accommodation srvice and controller test (#14)
* implemented accommodation srvice and controller test

* fixed after mvn clean package

* fixed after mvn clean package

* fixed

* fixed

* fixed

+++ src/main/java/booking/app/service/accommodation/address/AddressServiceImpl.java
@@ -40,6 +40,9 @@ public class AddressServiceImpl implements AddressService {
 
     @Override
     public void deleteById(final Long id) {
+        if (!addressRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't find address by id: " + id);
+        }
         addressRepository.deleteById(id);
     }
 }
